requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
crcmod==1.7
docopt==0.6.2
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
numpy==1.21.5
pandas==1.1.5
Pillow==9.0.1
pipreqs==0.4.11
pycparser==2.21
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
yarg==0.1.9
zipp==3.7.0
